update for directory build
