+++ .goreleaser.yml
@@ -4,7 +4,7 @@ before:
 builds:
   - env:
       - CGO_ENABLED=0
-    main: ./cmd/s3s/main.go
+    main: ./cmd/s3s
     ldflags:
       - -s -w
       - -X main.Version=v{{.Version}}
